refactor: isolate legacy walletconnectClient setup into own util
