+++ wallets/react-wallet-v2/src/utils/WalletConnectUtil.ts
@@ -1,9 +1,6 @@
 import SignClient from '@walletconnect/sign-client'
-import LegacySignClient from '@walletconnect/client'
-import ModalStore from '@/store/ModalStore'
 
 export let signClient: SignClient
-export let legacySignClient: LegacySignClient
 
 export async function createSignClient(relayerRegionURL: string) {
   signClient = await SignClient.init({
@@ -16,40 +13,5 @@ export async function createSignClient(relayerRegionURL: string) {
       url: 'https://walletconnect.com/',
       icons: ['https://avatars.githubusercontent.com/u/37784886']
     }
-  })
-}
-
-export function createLegacySignClient({ uri }: { uri: string }) {
-  legacySignClient = new LegacySignClient({ uri })
-  console.log('legacySignClient: ', legacySignClient)
-
-  console.log('BIND LEGACY LISTENERS')
-
-  legacySignClient.on('session_request', (error, payload) => {
-    if (error) {
-      throw new Error(`legacySignClient > session_request failed: ${error}`)
-    }
-
-    console.log('legacySignClient > session_request:', payload)
-    ModalStore.open('LegacySessionProposalModal', { legacyProposal: payload })
-  })
-
-  legacySignClient.on('connect', () => {
-    console.log('legacySignClient > connect')
-  })
-
-  // legacySignClient.on('disconnect', () => disconnect())
-
-  legacySignClient.on('error', error => {
-    throw new Error(`legacySignClient > on error: ${error}`)
-  })
-  legacySignClient.on('call_request', (error, payload) => {
-    if (error) {
-      throw new Error(`legacySignClient > call_request failed: ${error}`)
-    }
-
-    console.log('legacySignClient > call_request: ', payload)
-
-    // handleCallRequest(payload)
   })
 }
