+++ wallets/react-wallet-v2/src/views/LegacySessionProposalModal.tsx
@@ -4,7 +4,7 @@ import RequestModalContainer from '@/components/RequestModalContainer'
 import ModalStore from '@/store/ModalStore'
 import { eip155Addresses } from '@/utils/EIP155WalletUtil'
 import { isEIP155Chain } from '@/utils/HelperUtil'
-import { legacySignClient } from '@/utils/WalletConnectUtil'
+import { legacySignClient } from '@/utils/LegacyWalletConnectUtil'
 import { Button, Divider, Modal, Text } from '@nextui-org/react'
 import { getSdkError } from '@walletconnect/utils'
 import { Fragment, useState } from 'react'
